Scene Tree Cleanup
Cleaned up the scene tree by adding an entities layer (so far for the player and enemies), and a foreground layer (so far just the sword ability). This was done using groups in Godot. Additionally, this allowed for y sorting of the entities layer, which is great for creating depth perception in  top down games.

Did some other clean up by making types explicit.