Implementing player vignette shader
added a player vignette shader which is responsible for playing a PlayerDamaged animation when the player is hit. I also figured out where a bug is coming from when resetting the game, it looks to be related to game events not being instantiated again once the game has been reset, but will need to look into this more closely.