Sounds Effects (SFX) implemented!
got basic sound effects implemented for enemies being hit. For some unknown reason to man kind, this disabled enemy movement by changing the enemies max speed to 0... watch for this in the future. Additionally, I started a scene for the main menu but this isn't even close to done yet!