Implementing music
Got music implemented, note that this is not my music as of now, but is the mp3's used are licensed for free distribution. This is actually the case for all of the current sounds effects too. They are all currently place holders for music and SFX to come, although I might keep some of the SFX that I like such as: enemy damaged, clicks, and abilities cards.