Improving meta upgrade card
This commit contains A LOT of stuff tbh. The main thing was implementing the back end to apply an experience drop upgrade. This commit also included integrating the UI for the meta upgrade screen, which is why so many files have been changed. Also make a function for screen transitions because this code was being repeated a lot!