Getting game export ready
The game is currently export ready with some stub sprites for icons, awesome! Also, I looked into some of the bugs that are still present (namely the vignette, experience after restarting, and like ability upgrades), and it's totally related to the GameEvents signals in autoload. Will have to do more digging, but at least this was confirmed.